+++ src/funcs/getFolders.ts
@@ -18,6 +18,8 @@ export function getFolders(
     }
   }
 
+  console.log(reversedFolderPath);
+
   reversedFolderPath.reverse().forEach((folder) => {
     activeWorkspacePath += `\\${folder}`;
     folderPathsList.add(vscode.Uri.file(activeWorkspacePath));